refactor!: move Suit to separate module
This makes `lib.rs` smaller and easier to maintain.
