+++ src/lib.rs
@@ -1,11 +1,12 @@
 pub mod calc;
 pub mod fu;
 pub mod hand;
+pub mod suit;
 pub mod tile_group;
 pub mod yaku;
 
 use crate::fu::{calculate_total_fu_value, Fu};
-use crate::hand::error::HandErr;
+use crate::suit::Suit;
 use crate::tile_group::{GroupType, TileGroup};
 
 /// Characters that represent terminal or honor tiles.
@@ -19,41 +20,6 @@ pub enum LimitHands {
     Baiman,
     Sanbaiman,
     KazoeYakuman,
-}
-
-#[derive(Debug, Clone, PartialEq, Hash, Eq)]
-pub enum Suit {
-    Manzu,
-    Pinzu,
-    Souzu,
-    Wind,
-    Dragon,
-}
-
-impl Suit {
-    /// Parse the suit from the string.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use mahc::Suit;
-    ///
-    /// let tile_string = "9m";
-    /// let actual_suit = Suit::suit_from_string(tile_string.chars().nth(1).unwrap().to_string());
-    /// let expected = Ok(Suit::Manzu);
-    ///
-    /// assert_eq!(actual_suit, expected);
-    /// ```
-    pub fn suit_from_string(suit: String) -> Result<Self, HandErr> {
-        match suit.as_str() {
-            "s" => Ok(Self::Souzu),
-            "p" => Ok(Self::Pinzu),
-            "m" => Ok(Self::Manzu),
-            "w" => Ok(Self::Wind),
-            "d" => Ok(Self::Dragon),
-            _ => Err(HandErr::InvalidSuit),
-        }
-    }
 }
 
 impl LimitHands {
