+++ src/suit.rs
@@ -1,0 +1,36 @@
+use crate::hand::error::HandErr;
+
+#[derive(Debug, Clone, PartialEq, Hash, Eq)]
+pub enum Suit {
+    Manzu,
+    Pinzu,
+    Souzu,
+    Wind,
+    Dragon,
+}
+
+impl Suit {
+    /// Parse the suit from the string.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use mahc::suit::Suit;
+    ///
+    /// let tile_string = "9m";
+    /// let actual_suit = Suit::suit_from_string(tile_string.chars().nth(1).unwrap().to_string());
+    /// let expected = Ok(Suit::Manzu);
+    ///
+    /// assert_eq!(actual_suit, expected);
+    /// ```
+    pub fn suit_from_string(suit: String) -> Result<Self, HandErr> {
+        match suit.as_str() {
+            "s" => Ok(Self::Souzu),
+            "p" => Ok(Self::Pinzu),
+            "m" => Ok(Self::Manzu),
+            "w" => Ok(Self::Wind),
+            "d" => Ok(Self::Dragon),
+            _ => Err(HandErr::InvalidSuit),
+        }
+    }
+}
